Consolidate common heading tag code
Before: Heading[One|Two]TagElement extended TextTagElement
After:  New HeadingTagElement extends TextTagElement
        It is extended by the Heading[One|Two]TagElement classes
        They now consist soley of one-line constructors that call their super with their scale as an added parameter.
